Transfer Function related functions
The TransferFunction.py file is an updated version of the ARX.py file.
Time_MKS_Examples is a file used to displace the use of the functions
built in TransferFunction.py. I still need to work on developing
functions that use the coefficients of the model correctly. I seem to
get inaccurate answers.