Properties attached to quadrature points
Address #49

Material properties are now attached to quadrature points. Allows
removal of nasty interpolation of properties from elements to
vertices. This scheme fits better with the MKS and now allows the MKS
to accurately predict strains.

There is still an issue with the FFT in the x-direction due to lack of
periodicity in FE model.