Add MKSKernelPCAModel and make_microstructure
Address issue #80

Add MKSKernelPCAModel. Model takes in a microtructures with dimensions of
(n_samples, Nx, Ny,...) computes spatial correlations and does kernel
principle analysis on the correlations.

Add make_microstructure to provide a convenient interface for
microstructureGenerator.

The microstructure generator is currently partitioning an equal number of
bins for what looks like a normal distribution. It should equally partition
the cumulative distribution. A MKSKeneralPCAModel example still needs to be
made.