Fix broken Legendre polynomial test case
Address #83

The previous commit fixed the broken scaling for the Legendre
Polynomial bassis. However, the test case, which worked with the
broken scaling was not updated to work with the new (correct) scaling.
Furthermore the tests have been changed to explicitly calculate the
test comparison values by constructing the Legendre Polynomials in the
test.