CH simulation can use odd grid point count.
Issue #84

Previously the Cahn-Hilliard simulation was broken for an odd grid
point count. The code correctly assembles the k array so that even
arrays are (0, 1, -2, -1) for N=4 and odd arrays are (0, 1, 2, -2, -1)
for N=5. This has not been rigorously tested yet and a test case is
required as well as changing the notebooks to use odd N counts.