Change initial validation microstructures
Address issue #137

Change initial validation microstructures to be generated from an
uniform distribution to a normal distribution to better represent the
physical phenomenon.

The make_cahn_hilliard simulation remains unchanged. The initial
microstructures used for calibration need to be generated from an
uniform distribution in order to calibrate the simulation over the
entire local state domain.