Update non-periodic tests
Address issue #135

Update tests for non-periodic and periodic statistics. Correlation
now returns a real value as long as the imaginary component is less
the 1e-13.

This will allow the example to demonstrate both periodic and
non-periodic statistics on a checkerboard microstructure.

The tolerance keyword argument used in numpy's real_if_close function
may need to be adjusted for larger datasets.