Correct statistics calculation and tests
Address issue #148

Correct statistics calculation to only compute statistics for data
weighted by the confidence interval shown in the uncertainty mask.
The corresponding tests were also updated.

The current version of the code is returning the wrong number of
cross correlations for a 2 local state system. This needs to be
fixed still.