Update draw_microstructures function
Address issue #152

Update draw_microstructures function. The function would only take in
arrays that were 2D or a list of 2D arrays. Now the function can take
in a 3D array with n_samples as the first dimension of the array.

All of the examples that use this function have been updated.