Mean center data and update tests
Address issue #152

Mean center the data before the dimensionality reduction is done and
update doctests. This change is needed to get accurate results with
SVD.

If n_components is None, then n_components will be set to the
n_components value for the dimension_reducer. This change will reduce
confusion when passing a dimension reduction class.