Add check_module_installation
address issue# 233

Add check_module_installation function to verify that module is
installed and can be loaded. This change allows for the tests unittests
that rely on sfepy to be skipped if sfepy is installed without creating
errors in the test summary.

This change will add an almost negligible overhead to the tests runtime.

The doctests that rely on sfepy still need to be fixed.