Switch from nose to pytest
address #233

Switch from nose to pytest. The maintainer of nose has stepped down, while
pytest still has an active development community. Also pytest has
additional functionality over nose with options to import a module
or skip the test.

The current commit passes all the tests when sfepy is installed, and
skips all the tests related to sfepy when it is not installed. But, the
make function in datasets fail to import when sfepy is not installed.