Plot Principal Components vs Time
address #TBD

This change to tools.py creates a new function,
_draw_components_2D_time that creates 4 plots (iso - PC1,PC2,time;
PC2 vs. PC1; PC1 vs. time; PC2 vs. time). This is for when the
data follows a time series (as in the MD example). This creates
the plots to see if the components are correlated with time at all.
This should not have any side-effects from normal functionality.
If the user says "time=True" then the dataset for draw_components
should also contain a time series. If the user is not looking to
correlate with time then the functionality should be the same as
before.