updated to run tests without sfepy
addresses #233

this is needed to remove sfepy dependency for the tests

moves ElasticFESimulation import to local imports so one can import
pymks.datasets without sfepy. uses mock to setup and takedown a fixture that
redirects calls to pymks.datasets to pymks.simple_datasets by inserting a key:value
pair into sys.module when running pytests. addition of lines importing the functions
in the doctests is needed to reassign the function to the testing function in
pymks.simple_datasets. Otherwise the test will use the functions contained
in the file. the addition of a main routine serves to troubleshoot the test
fixtures

as a side effect, the doctests will never use sfepy now