Switch pyFFTW builders
address #232

Switch from using pyFFTW's interfaces to builders. This change allows
for the data to not be copied, have the input data be overwritten and
allows for computation to be computed in parallel. There was a strange
error that would occur the first time one of the builders was called
when planning_effort was set equal to 'FFTW_MEASURE'. Setting
planning_effort to 'FFTW_ESTIMATE' does decrease the performance
for FFTW when writing to disk, but it is still much faster that the
standard numpy.fft. See FFTW documentation on wisdom for more details.

As a result of switching to the builders, all of the classes that
used FFTs now inherit the the Filter class. Additionally some of the
internal attributes were renames with _ to meet PEP8 standards.