Use rFFTs instead of FFTs
address #232

Use rFFTs instead of FFTs when the input signal is real. This change
will allow for half of the memory to be saved using rFFTs. All of the
test currently pass, but there seems to be an issue when viewing the
coefficients after resizing them for certain sizes. Still need to
explore possible numerical instabilities with rFFTs.