Select correlations by index.
address #287
address #285

Select correlations by index. This change allows users to select which
correlations would like to selecting the index in n_states. As an
example if basis.n_states=[1, 3, 5] then correlations=[(0, 0), (1, 2)]
would compute the autocorrelation with 1 1 and the cross-correlation
with 3 and 5.

This change requires the users to know the indices of basis.n_states.