Switch to py.test and skip sfepy tests
Address #233

Switch to py.test as it allows an easy path to skip sfepy tests in the
event that the package is not installed. There is no easy way to skip
doctests that are dependent on sfepy other than having py.test skip an
entire module, thus, some tests have been removed from doctest and
made into standard `test_` functions.