Fix bug with microstructure generator
address #310

Prior to this change the range of volume fractions of microstructures
when using the argument percent_variance would occasionally be random.
With this change the variation in the microstructure is now match the
inputs provided.

The tests were also updated to something more general rather than
relying on numerical values.