Change seed to array
address #310

Update the seed in make_elastic_stress_random to include be an array
rather than an integer. Prior to this change,
make_elastic_stress_random what using the same seed to generate each
of the classes of microstructures. With this change, the seed is used
to create an array of seed values that are used for each class of
microstructure and increasing the variance in the data.

The tests were also updated to pass with the new seeds.