Update draw_microstructures and draw_gridscores_matrix
address #312

Update `draw_microstuctures` to take the argument `figsize` in order
to specify the number of microstructures plotted on each row and column.
Prior to this change all microstructures were plotted in one row.

Update `draw_gridscores_matrix` to have the two plots appear on top of
one another instead of side by side. Additionally I looked at plotting
the log of the errors, but found that clarity of the plots did not
improve.