Copy files from ./ into doc/ in the conf.py
Address #145

Copy the markdown files into doc in the conf.py so that they are
converted natively by Sphinx. Sphinx can not use the include directive
with markdown, which means that they need to be within the doc/
directory. This is an ongoing effort to make the documentation render
on Readthedocs.