Make Numpy fft default and explicilty call PyFFTW
Address #325

Make Numpy's fft module the default fft module. PyFFTW is used if
either an option is set in setup.cfg or an environment variable is
set. This avoids many issues where PyFFTW is installed, but breaks
things. There is no way to get back to using Numpy's fft without
uninstalling PyFFTW, which is a very complicated remedy for new
users. Since Numpy's fft is far more stable, it seems a better choice
just to always use it unless the user wants to try and use PyFFTW.

 - The fft functions are wrapped in a better manner than previously
   implemented to reduce code duplication.