Remove the ImportError qualification for pyfftw
Address #325

Remove the ImportError qualification for the try and except block for
pyfftw. If there are any issues importing pyfftw that aren't an
ImportError than we don't want failures for that. We want it to fail
with the correct error in fftmodule.py when the real import is
executed, not on the import to just get round the (numpy, pyfftw)
import order bug.