Fix broken use of __name__ in setup.py
Setup.py had been changed to not import pymks requiring the
get_version function functionality to be used in setup.py. This
referred to __name__, which is "pymks" in __init__.py, but in setup.py
is __main__. This broke both the install when git wasn't present and
the conda build.