Switch Numpy to using blas from openblas
The combination of Scikit-learn / Numpy / Openblas seems to be
returning strange values for this version of Nixpkgs. The fiber.ipynb
and stress.ipynb notebooks get very strange results. This seems to be
related to a dot product in Numpy when doing a PCA with
Scikit-learn. This fixed itself when using the standard Blas in
Nixpkgs. All tests now pass.