Fix bugs when using Ubuntu 19.04 (#390)
Fix for an issue when using Dask and Ubuntu 19.04 with the versions specified in the Nix installation. This bug occurred with distributed.Client(), resulting in an error inside psutil to do with reading files in the the latest version of Ubuntu (19.04). Updating to Nixpkgs version 18.09 resolved this issue, but resulted in another issue with the interaction of Numpy and the openblas library. To resolve that, Numpy was switched to using the blas package in Nixpkgs instead of the openblas package.

 * Updated version of distributed to avoid psutil bug

 * Use filterSource to avoid including .git directory in Nix
   installation which uses a lot of memory

 * Fix MPI bug when using Sfepy by adding ssh path

 * Update version of distributed to fix psutil bug

 * Change version of pytest-cov to make compatible with pytest

 * Add new regex in sanitize.cfg to avoid Matplotlib failures in notebook tests with newer version of Matplotlib

* Switch Numpy to using blas from openblas to avoid weird np.dot restult that looked like some sort of memory issue with repeated values in the numpy array.

* Remove references to sklearn.nix as just using standard Nix version of sklearn now

* Fix deprecation warning when using a list to index a Numpy array