Implement functional version of Sfepy module (#395)
Sfepy module is now programmed in a functional way. It works with both Numpy and Dask arrays. The Sfepy module has a `solve` function that returns a dictionary with "stress", "strain" and "displacement" keys.

Some of the notebooks only work intermittently. It seems like there is currently an issue with Scikit-learn / numpy / blas configuration in the Nix build. This was a problem in recent commits on develop as well. It was thought that this had been fixed by changing the blas version for Numpy, but a similar issue is popping up again. It seems like it is to do with PCA returning strange results in Scikit-learn.