build(conda): use more recent package versions
Address #414

Update Conda recipe to use more recent version of packages to more
closely match package versions in the Nix recipe. Tests pass with both
Nix and Conda.

Rename Conda environment to pymks instead of conda-test1 as pymks is a
more descriptive name of what the environment is.

Note that still using Sfepy version 2018.2 and Python 3.6 as Sfepy
won't build in Conda with Python version 3.7 with any version of Sfepy
and won't build 2019.2 with Python 3.6.