build(conda): remove version numbers from build
Assigning version numbers to each of the packages in the Conda build
results in consistent failures of the Conda package solver. Conda with
conda-forge will only work for a short time with assigned packages. As
soon as packages are updated then Conda finds conditions that can not
be unresolved. This does have the benefit of keeping PyMKS up to date
with the latest packages as development moves forward.

Keep matplotlib at versin 3.1 for compatibility with Nix