build(nix): fix development environment
The development environment was broken when using nix-shell. The
version of pymks used was from the nix store not from the local
working repository unless running python from withing the local base
working repository. This showed up when running a notebook. The
imported pymks was the nix store installed version at the time of
running nix-shell and didn't include any edits to the working
copy. Using filterSource to filter the source tree was doing something
so that buildPythonPackage didn't recognize that the pymks was a
development environment. cleanSource seems to avoid that problem.